Feature

+++ Camel.java
@@ -2,7 +2,7 @@ package Animals.Animals;
 
 public class Camel extends Pack_animals {
 
-    public Camel(String name, String dateOfBirth, Float carryingCapacity) {
+    public Camel(String name, String dateOfBirth, Double carryingCapacity) {
         super(name, dateOfBirth, carryingCapacity);
     }
